// packages/playwright-tools/types.d.ts
/**
 * Copyright (c) Microsoft Corporation.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import type playwright from 'playwright';

export type JSONSchemaType = string | number | boolean | JSONSchemaObject | JSONSchemaArray | null;
interface JSONSchemaObject { [key: string]: JSONSchemaType; }
interface JSONSchemaArray extends Array<JSONSchemaType> {}

export type ToolDeclaration = {
  name: string;
  description: string;
  parameters: any;
};